Get rid of ShiftObliviousSwapListener
