+++ engine/table/src/main/java/io/deephaven/engine/table/impl/sources/ring/RingTableTools.java
@@ -8,7 +8,6 @@ import io.deephaven.engine.table.TableUpdate;
 import io.deephaven.engine.table.impl.BaseTable;
 import io.deephaven.engine.table.impl.SwapListener;
 import io.deephaven.engine.table.impl.perf.QueryPerformanceRecorder;
-import io.deephaven.engine.table.impl.remote.ConstructSnapshot;
 import io.deephaven.engine.table.impl.remote.ConstructSnapshot.SnapshotFunction;
 import io.deephaven.engine.table.impl.sources.ring.AddsToRingsListener.Init;
